% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{collect_metadata}
\alias{collect_metadata}
\title{Collect meta data for each storify from a vector of Storify URLs.}
\usage{
collect_metadata(vector_of_urls)
}
\arguments{
\item{vector_of_urls}{A vector of Storify URLs.}
}
\value{
out_df A data frame with the number of rows equal to the length of `vector_of_urls` with metadata for each Storify.
}
\description{
Collect meta data for each storify from a vector of Storify URLs.
}

