% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scrape_data}
\alias{scrape_data}
\title{Scrape data for every tweet in every Storify for which metadata was collected}
\usage{
scrape_data(storify_metadata, output_folder = NULL)
}
\arguments{
\item{storify_metadata}{A data frame with a URL for each Storify and associated metadata as output from `collect_metadata()`.}

\item{output_folder}{Folder for the data frames for each Storify written as a CSV.}
}
\value{
out_list A list with the same length as the number number of rows in `storify_metadata`, with each list item a data frame with data for all of the tweets scraped. In addition, each data frame is written to a CSV file in the `output_folder` argument.
}
\description{
Scrape data for every tweet in every Storify for which metadata was collected
}

